feat: logger for why reconnection is attempted
